Move all models into \App namespace and cleanup unused use statements
