+++ app/Conference.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App;
 
 use Auth;
 use Carbon\Carbon;
