+++ app/Events/ProfilePictureUpdated.php
@@ -2,12 +2,8 @@
 
 namespace App\Events;
 
-use App\Events\Event;
-use App\Models\User;
-use Illuminate\Http\Request;
-use Illuminate\Http\UploadedFile;
+use App\User;
 use Illuminate\Queue\SerializesModels;
-use Illuminate\Contracts\Broadcasting\ShouldBroadcast;
 
 class ProfilePictureUpdated extends Event
 {
