+++ app/Http/Controllers/AccountController.php
@@ -1,13 +1,12 @@
 <?php namespace App\Http\Controllers;
 
 use App\Events\ProfilePictureUpdated;
-use App\Models\User;
+use App\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Event;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Session;
-use Intervention\Image\Facades\Image;
 
 class AccountController extends BaseController
 {
