+++ app/Http/Controllers/Api/BiosController.php
@@ -1,11 +1,9 @@
 <?php namespace App\Http\Controllers\Api;
 
 use App;
-use Exception;
-use Illuminate\Support\Facades\Auth;
 use App\ApiResources\Bio;
 use App\OAuthGuard\Facades\OAuthGuard;
-use App\Models\User;
+use Exception;
 
 class BiosController extends BaseController
 {
