+++ app/Http/Controllers/Api/UserBiosController.php
@@ -1,9 +1,8 @@
 <?php namespace App\Http\Controllers\Api;
 
-use Illuminate\Support\Facades\App;
 use App\ApiResources\Bio;
 use App\OAuthGuard\Facades\OAuthGuard;
-use App\Models\User;
+use Illuminate\Support\Facades\App;
 
 class UserBiosController extends BaseController
 {
