+++ app/Http/Controllers/PublicProfileController.php
@@ -2,13 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\User;
 use Captcha\Captcha;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Mail;
 use Illuminate\Support\Facades\Session;
-use App\Http\Controllers\Controller;
-use App\Models\User;
 
 class PublicProfileController extends Controller
 {
