+++ app/Listeners/UpdateProfilePicture.php
@@ -3,10 +3,7 @@
 namespace App\Listeners;
 
 use App\Events\ProfilePictureUpdated;
-use App\Models\User;
-use Illuminate\Contracts\Queue\ShouldQueue;
-use Illuminate\Queue\InteractsWithQueue;
-use Illuminate\Support\Facades\File;
+use App\User;
 use Illuminate\Support\Facades\Storage;
 use Intervention\Image\Facades\Image;
