+++ app/Services/Registrar.php
@@ -1,8 +1,8 @@
 <?php namespace App\Services;
 
-use App\Models\User;
-use Validator;
+use App\User;
 use Illuminate\Contracts\Auth\Registrar as RegistrarContract;
+use Validator;
 
 class Registrar implements RegistrarContract
 {
