+++ database/seeds/BiosSeeder.php
@@ -1,7 +1,7 @@
 <?php
 
-use App\Models\Bio;
-use App\Models\User;
+use App\Bio;
+use App\User;
 use Illuminate\Database\Seeder;
 
 class BiosSeeder extends Seeder
