+++ database/seeds/UsersSeeder.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Models\User;
+use App\User;
 use Faker\Factory as Faker;
 use Illuminate\Database\Seeder;
 use Laracasts\TestDummy\Factory;
