+++ tests/Api/BioApiTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Models\Bio;
+use App\Bio;
 use Illuminate\Foundation\Testing\WithoutMiddleware;
 
 class BioApiTest extends ApiTestCase
