+++ tests/Api/ConferenceApiTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Models\Conference;
+use App\Conference;
 use Illuminate\Foundation\Testing\WithoutMiddleware;
 
 class ConferenceApiTest extends ApiTestCase
